Move HeaderRenderer out of components/common.js
